test(frontend): align settings helper coverage
