+++ packages/frontend/src/components/features/settings/settings-prompt-inheritance.test.ts
@@ -1,0 +1,54 @@
+import { describe, expect, test } from "bun:test";
+import { buildInheritedPromptPreview } from "./settings-prompt-inheritance";
+
+describe("settings prompt inheritance", () => {
+  test("builds inherited preview from global override and builtin", () => {
+    const fromGlobal = buildInheritedPromptPreview(
+      "kickoff.spec_initial",
+      {
+        template: "repo override",
+        baseVersion: 2,
+        enabled: false,
+      },
+      {
+        "kickoff.spec_initial": {
+          template: "global override",
+          baseVersion: 1,
+          enabled: true,
+        },
+      },
+      "builtin",
+    );
+    expect(fromGlobal).toEqual({
+      sourceLabel: "Global override",
+      template: "global override",
+    });
+
+    const fromBuiltin = buildInheritedPromptPreview(
+      "kickoff.spec_initial",
+      {
+        template: "repo override",
+        baseVersion: 2,
+        enabled: false,
+      },
+      {},
+      "builtin prompt",
+    );
+    expect(fromBuiltin).toEqual({
+      sourceLabel: "Builtin prompt",
+      template: "builtin prompt",
+    });
+
+    const hiddenWhenRepoEnabled = buildInheritedPromptPreview(
+      "kickoff.spec_initial",
+      {
+        template: "repo override",
+        baseVersion: 2,
+        enabled: true,
+      },
+      {},
+      "builtin prompt",
+    );
+    expect(hiddenWhenRepoEnabled).toBeUndefined();
+  });
+});
